refactor: organize files based on lecture
