+++ EssentialFeed/EssentialFeediOS/Feed Image Loader/FeedImageDataLoader.swift
@@ -1,0 +1,17 @@
+//
+//  FeedImageDataLoader.swift
+//  EssentialFeediOS
+//
+//  Created by Nikhil Menon on 1/30/24.
+//
+
+import Foundation
+
+public protocol FeedImageDataLoader {
+    typealias Result = Swift.Result<Data, Error>
+    func loadImageData(from url: URL, completion: @escaping ((FeedImageDataLoader.Result) -> Void)) -> FeedImageDataLoaderTask
+}
+
+public protocol FeedImageDataLoaderTask {
+    func cancel()
+}
